move svm light loader, add some comments noting this should usually not be used unless someone has a hankering for provable optimality
